+++ projects/swagger-schematics/helpers/config.ts
@@ -8,6 +8,7 @@ type swaggerSchemaKeys = keyof SwaggerApiSchema;
 /** Default configuration values */
 const DEFAULT_CONFIG: Partial<SwaggerApiSchema> = {
     scopeEndpointsWithTags: false,
+    eslintFix: false,
 };
 
 export const getOpenapiSchematicsConfig = (options: SwaggerApiSchema): SwaggerApiSchema => {